Allow to update an alarm partially
The patch allow to only modify a part of an alarm instead of force to
set the full alarm description.

This permit to an application that have been written code around alarm
with ceilometerclient 1.0.4 and ceilometer pre havana-3. To update alarm
without any change with this and ceilometer >= havana-3 (ie: heat).

Fixes bug #1231303

Change-Id: I20250131d05d20bfadbca450dfe6b8237f4b7183