align the order of parameters for urlencode()
In Python 3.3, hash randomization is enabled by default. It causes the
iteration order of dicts and sets to be unpredictable and differ
across Python runs.

In the test case, the fixed expecting string will not match the test
result, it is relying on the dict order.

This change transforms the input dict to a sequence of two-element list,
with fixed order, and update the related expecitng string / fixture
in test cases.

Partial Implement: blueprint py33-support

Change-Id: I6dccde9e584be8335a6375f5fbad5c5cbd7b9b6d