Checking the type of auth_token, and fixing it if necessary
When creating a new Client object by calling the get_client function,
if the ceilometer_url is not passed as a parameter, but the token is
passed, the auth_token is converted to a lambda function.
However, when both the ceilometer_url and the auth_token are passed,
this conversion does not take place.
This fix checks if the auth_token is a lambda function or not.
If it is not, it creates a new lambda function which returns the
token.
Without this check, if a token was passed as a string, the
python client would throw and error, when trying to call
self.auth_token() in ceilomenterclient/common/http.py.

Change-Id: I3e564fcf17ce497546b75f0ab8c7ca0f6f2099df
Closes-Bug: #1254186