add notification alarm
allow alarms to be created based on a notification type or a
payload of a notification. For example: you can create an
alarm using the following command:

ceilometer alarm-notification-create
     --name network-alarm
     --notification-type "network.*"
     -q "network.name=example;age<5"

this will create an alarm, when a notification type matches
network.* and the payload of the notification with network.name
equals word "example", and the age less than 5, an alarm will
be triggered.

client code changes for the blueprint.

Change-Id: If62c32162937ce2fc496b42bbd10af70c2d20bf1
bp: https://blueprints.launchpad.net/ceilometer/+spec/alarm-on-notification