Use Emacs-friendly file variable to set file encoding
The file encoding line serves a dual purpose: it informs both Python
and Emacs about the file encoding.

PEP 263 explains that a file with non-ASCII characters must specify an
encoding. Without an encoding line, importing the file will raise a
SyntaxError.

The problem is now that while Python will accept a line with

  -*- encoding: <some-encoding> -*-

Emacs only accepts

  -*- coding: <some-encoding> -*-

since it looks for a file variable named "coding", not "encoding".

Setting a variable called "encoding" has at best no effect, but it can
result in Emacs prompting the user with

  The local variable list in __init__.py contains values that may not
  be safe.

when opening the file. This is because the "encoding" variable is not
whitelisted as a safe variable.

The command used to update the files were:

  sed -i '1,2s/-\*- encoding:/-*- coding:/' **/*.py

This accounts for the case where the file variables is set in the
second line because the first line is used for a "#!" line.

Change-Id: I6b43cc5ecfd0440e06b0aebd91ebb574b7f37a91