Refactor split_by_op for greater encapsulation
split_by_op was returning data against which further error checking was
then required. This change moves that error handling inside the method
and rationalizes the regular expressions used so they are only doing
the split operation, not constraining what is being split.

Operators are allowed in values, but not keys.

Tests added to cover the method more completely. 'string' variable
renamed to 'query' to avoid confusion. Named parameters on string
substitution for clarity.

Change-Id: Icd815ff65aba9eae3f76afee3bb33e85d85bea72