Fix the alarm history order shown to user
In the Ceilometer storage backends implementation we are specially
sorting the alarm history points in the following way:
the first one by the timestamp should be the last one in the list.
The Python client itself provides the information in this way, but
the table printed had the opposite one, so this was fixed.

Change-Id: If39bb5868e360f37ed76c12b55fc71eade331f47