Alarm TimeConstraint display incorrect
This patch fixes the incorrect display of alarm timeconstraint.
By avoiding the late sorting of the display dict on property and
sorting the dict eagerly/early, we avoid the issue of
timeconstraint rows appearing at the top of the display table.
By wrapping the lines after they split, the spurious newline row
is avoided.

Change-Id: I3ab8a4e25f946b409ede2a598a5b9cf8593a044e
Closes-Bug: #1395110