add --no-traits for event-list
Currently we print traits list for each event when run event-list
command, but that is a lot of information. After system run for a
few days, the event-list command prints so much that it is not
convenient to know the whole event-list even with less command.

This patch adds a simple argument called --no-traits, it is False
by default, event-list will not print traits if this argument is
specified. Since this argument is boolean type, no need to assign
a value, if it is specified, then it is actived.

Change-Id: Ib69209fef5967a18f094f2d4665fa189e4ca3e6a