Add option to print specific fields from an Event
Currently the only way for the client to print an event is to print the
fields in a box-like structure in the command line. This is fine for
many use-cases, but if we want to automate things in the command line,
or pipe some specific field from the event (such as the Event's traits)
then being able to specify which field will be exclusively printed
becomes very useful.

Here's an example:

ceilometer event-show $EVENT_ID -f traits | python -m json.tool

The following would exclusively print the Event's traits (which is
already in JSON format) and the output could be nicely printed in the
command line for further inspection.

Change-Id: I5b9b8fae88a4cef01a4e42513e3b02c633e3c6e2