Catch missing 404 exceptions in client
Whenever an entity wasn't found, the ceilometerclient would issue a
NotFound exception. Which is fine, except that it's not the exceptions
that was being excepted in the client at all. This exception was defined
in the apiclient in the openstack/common directory. So, instead this
exception is now being caught in the manager, and now the appropriate
exception is being raised.

There were some functions where, if the entity was not to be found, a
404 exception would be raised, which was never caught by the client
since it was expecting the function to return None, instead of raising.
This CR fixes that.

Closes-Bug: #1451833
Change-Id: I47b34af2df9c85f86ba1f7f4d7b2951f8d27f96c