Fix requirements.txt to be compatible with dh_python2
The current requirements.txt has two dependencies (pbr and
requests) that have version exclusion clauses following their
names. This is causing the dh_python2 script, which is used in
Ubuntu packaging python-ceilometerclient for 2015.1.1, to fail.
The reason is due to the way that dh_python2 performs the parsing
for the requirements. The workaround fix is to move the exclusion
clauses to the end of the line.

Change-Id: I7078da3d1f198f4d95770c29de362d56260f6c8d