change the dict output format to make consistency under py27 and py34
For the command output in ceilometer cli, if it is a dict and it has
items which is list type, originally it won't handle the items in the
list. 

As to under py27, the string type in the list is unicode, the string
always look like u"http://alarm", while under py34 the string is
"http://alarm", this cause output inconsistency. This patch will handle
this inconsistency.

Change-Id: I5e7cf4052f68e5a434bcf86242b11365b34750a0