Enhances client to support unique meter retrieval
meter list API is enhanced to support a unique param
If specified as True, it returns a unique list of
meters.

This patch enhances the ceilometer-client to support
the unique argument in the shell and the client bindings.

DocImpact
Closes-Bug: #1506959

Change-Id: I011ca725c25e1cdbe9891c3719f13f5cd34142ea
(cherry picked from commit c45000c7594ad6fbfabf44c0615dce3021f43605)