Prints  '-' instead of 'None' when data is None
Client prints 'None' when data is None , and client also prints
'None' when data is a string 'None'. But string 'None' is  different
to None. To clear the confusion, Client should print '-' instead of
'None' when data is None. Nova Client and Cinder Client has cleared
the confusion.

Change-Id: I71a2d105cc49eefd8509eb6712b21c652fcb694d
Related-Bug: #1266414
Related-Bug: #1422244