[bugfix]with keystone v3 Could not find domain: default.
when we use adminrc with keystone v3  not set OS_PROJECT_DOMAIN_ID
OS_USER_DOAMIN_ID,default ceilometerclient code will set it to
string default.this will lead error.

value is often like fdc79c4b63e8492d98271412320e8672
so when change this code to set it as None.
last we delete the user_domain_id and project_domain_id
because it is not useful.

close-bug: 1679934

Change-Id: I95c994406ed7a0b15ba6446e80f01e4f6787ec20